\name{chop}
\alias{chop}
\title{Remove last characters from strings}
\description{This function removes the last "n" characters from a string or vector of strings.
}
\usage{
chop(s,n)
}

\arguments{
  \item{s}{a string or vector of strings}
  \item{n}{the number of characters to remove}
}

\value{
The input strings with the last n characters removed.
}
\author{Claudio Ortelli}
\keyword{ utilities}