\name{get.tclSortModes}
\alias{get.tclSortModes}
\title{Get the sortmodes for tcl types}
\description{
Given a vector x of tcl types, the functions returns a vector of
the same length as x with the corresponding sort modes.
}
\usage{
get.tclSortModes(x)
}

\arguments{
  \item{x}{a vector of tcl types}
}
\details{
da fare
}
\value{
A vector of the same length than x containing the corresponding sort modes.
}
\author{Claudio Ortelli}
\keyword{ utilities}
